refactor: loadingCircleの切り出し
