+++ components/shared/LoadingCircle/index.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css'
+
+export const loadingCircleContainer = style({
+  display: 'flex',
+  justifyContent: 'center',
+  padding: '20px 0',
+})
